only disable interrupts with lower priority than SDHC DMA in AudioPlaySdRaw::play()
