+++ Audio.h
@@ -55,6 +55,37 @@
 #define AudioNoInterrupts() (NVIC_DISABLE_IRQ(IRQ_SOFTWARE))
 #define AudioInterrupts()   (NVIC_ENABLE_IRQ(IRQ_SOFTWARE))
 
+/**
+  \brief Get Base Priority
+  \details Returns the current value of the Base Priority register
+  \return Base Priority register value
+ */
+static inline uint32_t __get_BASEPRI() __attribute__((always_inline));
+
+static inline uint32_t __get_BASEPRI() {
+	uint32_t result;
+	__asm volatile(
+		"mrs %0, basepri"
+		: "=r" (result)
+	);
+	return result;
+}
+
+
+/**
+  \brief Set Base Priority
+  \details Assigns the given value to the Base Priority register
+  \param[in] basePri Base Priority value to set
+ */
+static inline void __set_BASEPRI(const uint32_t value) __attribute__((always_inline));
+
+static inline void __set_BASEPRI(const uint32_t value) {
+	__asm volatile(
+		"msr basepri, %0"
+		:: "r" (value) : "memory"
+	);
+}
+
 // include all the library headers, so a sketch can use a single
 // #include <Audio.h> to get the whole library
 //
