+++ play_sd_wav.cpp
@@ -25,6 +25,7 @@
  */
 
 #include <Arduino.h>
+#include "Audio.h"
 #include "play_sd_wav.h"
 #include "spi_interrupt.h"
 
@@ -42,37 +43,6 @@
 #define STATE_PARSE4			11 // ignoring unknown chunk after "fmt "
 #define STATE_PARSE5			12 // ignoring unknown chunk before "fmt "
 #define STATE_STOP			13
-
-/**
-  \brief Get Base Priority
-  \details Returns the current value of the Base Priority register
-  \return Base Priority register value
- */
-static inline uint32_t __get_BASEPRI() __attribute__((always_inline));
- 
-static inline uint32_t __get_BASEPRI() {
-	uint32_t result;
-	__asm volatile(
-		"mrs %0, basepri"
-		: "=r" (result)
-	);
-	return result;
-}
-
-
-/**
-  \brief Set Base Priority
-  \details Assigns the given value to the Base Priority register
-  \param[in] basePri Base Priority value to set
- */
-static inline void __set_BASEPRI(const uint32_t value) __attribute__((always_inline));
-
-static inline void __set_BASEPRI(const uint32_t value) {
-	__asm volatile(
-		"msr basepri, %0"
-		:: "r" (value) : "memory"
-	);
-}
 
 void AudioPlaySdWav::begin(void)
 {
